ajustando creacion de productos y agregando mensajes de dialogo de eliminación

+++ src/app/components/products/product-list/product-list.component.ts
@@ -3,6 +3,7 @@ import { Product } from 'src/app/model/products';
 import { Subscription } from 'rxjs';
 import { ProductService } from 'src/app/services/product.service';
 import { ProductospruebaService } from 'src/app/services/productosprueba.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-list',
@@ -42,20 +43,41 @@ export class ProductListComponent implements OnInit, OnDestroy {
     this.productService.setAll(this.productospruebaService.getDummyProducts());
   }
 
-  deleteProduct(nombre: string): void {
-    this.productService.delete(nombre);
+  async deleteProduct(nombre: string) {
+    const result = await Swal.fire({
+      title: '¿Está seguro que desea eliminar el producto ?',
+      text: 'Esta acción no se puede deshacer.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'Cancelar',
+    });
+
+    if (result.isConfirmed) {
+      await this.productService.delete(nombre);
+      // Muestra un mensaje de confirmación de eliminación exitosa
+      Swal.fire('¡Eliminado!', 'El producto ha sido eliminado.', 'success');
+    } else if (result.dismiss === Swal.DismissReason.cancel) {
+      // Muestra un mensaje de cancelación
+      Swal.fire('Cancelado', 'El producto no ha sido eliminado.', 'error');
+    }
+
+
+
+
+
   }
   updateQuantity(product: Product): void {
     this.productService.update(this.productList.indexOf(product), product);
   }
 
-  originalProductList: Product[] = [];  
+  originalProductList: Product[] = [];
 
   filter(event:Event){
     const filterValue = (event.target as HTMLInputElement).value;
     if (filterValue) {
       // Filtra la lista de productos si hay un valor de filtro
-      this.productList = this.originalProductList.filter(product => 
+      this.productList = this.originalProductList.filter(product =>
         product.nombre.toLowerCase().includes(filterValue.toLowerCase()) ||
         product.precio <= Number(filterValue)
       );
